refactor(dropdown): create list item and list item title components
